refactor: extract DiscoveryRegistry from ChunkManager
Moves all discovery-state tracking (object IDs, mark/check discovery,
the getDiscoveredX family, region discoveries, and chunk-lookup helpers)
out of ChunkManager into a new DiscoveryRegistry class. ChunkManager
delegates through its existing public API, so no consumers change.

- ChunkManager: 2,845 -> 1,818 lines; now focused on chunk lifecycle
  and generation
- DiscoveryRegistry accesses chunks through a narrow ChunkSource
  interface, making it independently testable
- discoveredObjects/discoveredRegions remain accessible on ChunkManager
  via back-compat getters (used by SaveLoadService and debug-spawner)
- Adds 21 unit tests covering ID generation, discovery round-trips,
  reconstruction from stored data, and delegation back-compat
